Ensure that the custom model interpolator is fully initialized
Importing the JBoss Java EE 6 specs bom
(org.jboss.spec:jboss-javaee-6.0:1.0.0.Final) revealed a problem with
ProjectPropertiesModelInterpolator. It’s an indirect subclass of
AbstractStringBasedModelInterpolator which has a UrlNormalizer field
and a PathTranslator field. Both of these were being left with null
values. The null UrlNormalizer was causing an NPE during interpolation
of the JBoss bom’s model.

Both fields are annotated with Plexus’ @Requirement annotation. This
commit initialises both fields with the default implementations of
UrlNormalizer and PathTranslator and adds a test to verify that the
JBoss bom can be used.

Fixes #3