Ensure that project’s group and name are used in gradle-plugin property
The configure closure’s delegate is the task so group and name were
referencing the task’s group and the task’s name respectively. The task
has a reference to its project so this commit uses the project reference
to ensure that the project’s group and name are used.

See #4