Extending configurations should inherit dependency management
Previously, configuration-specific dependency management was isolated
to the configuration on which it was declared and any configurations
that extended this configuration (thereby inheriting its dependencies)
would not inherit the dependency management. This could lead to
the compile configuration having one version of a dependency and the
testRuntime configuration having another, even when the compile
configuration was the only place where the dependency was declared.

This commit updates the plugin so that dependency management is
inherited in the same way that dependencies are. This ensures that a
dependency will, by default, have the same version across a
configuration hierarchy unless the dependency or dependency management
has been declared at multiple places in the hierarchy with different
versions.

Fixes #7