Make system properties available to the Maven model builder
The Maven model builder may require access to the current system
properties. For example, JdkVersionProfileActivator uses the
java.version system property to determine whether or not a profile in a
pom should be activated. In the absence of a required system property
the model fails to build and its dependency management is unavailable.

This commit adds the current system properties to the model building
request, thereby making them available to the model building process.
Handling of model building failures has also been improved so that any
errors are logged. Arguably a model building error should cause the
build to fail but throwing an exception (the expected way of stopping
a Gradle build) does not stop the build when its thrown during
dependency resolution.

Fixes #18