Apply global dependency management to generated Maven poms
Any dependency management that's declared globally, i.e. it isn't
specific to an individual configuration, is now reflected in the
Maven poms generated by the maven and maven-publish plugins. Imported
boms are declared as dependencies with <type>pom</type> and
<scope>import</scope>. Individual managed dependencies or sets of
managed dependencies are declared as standard dependencies within the
pom's <dependencyManagement/>.

Closes #8
Closes #19