Don’t allow exclusions to affect extending configurations
Previously, when the plugin identified that an artifact was to be
excluded from a configuration, the exclusion was applied to the
configuration. In Gradle, when one configuration extends another
configuration, the extending configuration inherits the exclusions from
the configuration that’s being extended. Even if the extending
configuration had a direct dependency on the artifact it was still
excluded as an exclusion added to a configuration applies to all, both
direct and transitive.

This commit updates the plugin to add exclusions to dependencies rather
than the whole configuration. This prevents the exclusions from being
inherited by extending configurations.

Fixes gh-21