Relax validation of distributionManagement.status in Maven poms
Some pom files, for example those of Apache Axis, contain a <status> 
element in their <distributionManagement> section. This is considered
invalid by Maven’s DefaultModelValidator. The plugin isn’t affected by
such problems and, rather than logging lots of problems when a project
depends on an artifact with such a pom, we can safely ignore them.

This commit introduces a new RelaxedModelValidator that’s used when
building the Model from a pom. The relaxed validator checks for a
<status> element within the <distributionManagement> section and, if
it finds one, it removes it for the duration of the validation.

See gh-22