Batch pom artifact resolution where possible and cache results
Previously the plugin used a detached configuration per pom artifact
that it needed to resolve. Despite Gradle having its own caching, the
process of resolving a configuration that points to a single, cached
file was slow. Some experimentation has revealed that batching the
resolution of the pom artifacts into a single resolve performs
significantly better.

This commit updates ExclusionResolver to be driven with a set of
dependencies for which the poms are to be resolved. Once the poms have
been resolved and the exclusions determined, the result is cached to
minimise the processing required when the same dependency is encountered
again.

While it doesn’t use batching, PomDependencyModelResolver has also been
updated to cache the pom files that it has resolved. This minimises the
number of configuration resolutions that are performed.

See #22