Ignore repeated dependencies when processing graph to find exclusions
When exclusions are being found a particular dependency only needs to
be considered once but the same dependency may appear in the graph
multiple times. This commit updates the processing to only consider each
dependency once when visiting the graph to find exclusions. Note that,
when determining which exclusions to apply, every instance of each
dependency in the graph is considered so that any unexcluded edges can
be found and the exclusion ignored.