Add support for copying imported boms into a generated pom
Gradle's pom generation requires any boms referenced in a pom's
<dependencyManagement> section to be available from Maven Central and
publishing will fail if this is not the case. To work around this
limitation this commit adds support for configuring the plugin to copy
imported boms into the generated pom rather than importing them. It also
adds support for disabling the customisation of generated poms entirely
so that a user can take complete control over the process should they
so wish.

Closes gh-27