Add support for declaring dependency management exclusions
Previously, exclusions could only be declared using Gradle’s standard
mechanisms. One downside of this was that those exclusions would be
applied with Gradle’s semantics which are not the same as Maven’s. A
second downside is that they would not be predictably included in a 
generated pom – configuration-wide excludes are never included and
exclusions on a dependency are only included when they declare both a
group and a module.

This commit adds support for declaring exclusions using the plugin’s
DSL. The exclusions are applied using Maven’s semantics and are included
in a generated pom. Exclusions can be declared on individual
dependencies and on individual entries in a dependency set.

Closes #1