Tolerate dependencies with circular references
org.apache.xmlgraphics:batik-rasterizer as transitive dependencies that
are circularly dependent: org.apache.xmlgraphics:batik-svg-dom and
org.apache.xmlgraphics:batik-anim depend on each other. This cycle
caused a StackOverflowException when the Maven exclusion support was
building a model of the dependency graph.

This commit updates DependencyGraph so that it tolerates cyclic graphs
by keeping track of the modules that it has seen and short circuiting
the traversal when it encounters a module that has already been
processed.

Fixes gh-33