Align the dependencies syntax more closely with Gradle's DSL
Previously a dependency within the dependencies block was configured
using 'group:name' 'version'. This was at odds with Gradle's own DSL
where the equivalent syntax would be dependency 'group:name:version'.

This commit updates the plugin to align with Gradle's DSL. The old
syntax is supported, but deprecated. Using it produces a warning that
recommends migrating to the new syntax and advices the user that
support for the old syntax will be removed in a future release.

Closes gh-35