Automatically set the Spock version based on the Groovy version
Spock comes in two flavours; one that supports Groovy 1.8 and one that
supports Groovy 2.x. This commit updates build.gradle to calculate the
version of Spock to use based on the version of Groovy that’s being
used. This makes it easier to build the plugin against Gradle 1.x
(Groovy 1.8) and Gradle 2.x (Groovy 2.x). 

One test in DependencyManagementPluginSpec has been updated as part of
this commit. Gradle 2.x resolves 25 files (rather than the 24 resolved
with 1.x). I believe this is due to the different handling of 
relocated dependencies between the two versions.