Allow caching of imported boms to be configured
If a project using a snapshot bom, it is useful to be able to configure
the caching of the snapshots. Usually this would be done by
customising a configuration’s resolution strategy, however the plugin
does not expose the configurations that are used to resolve imported
boms.

This commit adds support for using a resolutionStrategy closure to
configure the resolution strategy of any configurations that are created
by the plugin. For example:

dependencyManagement {
    resolutionStrategy {
        cacheChangingModulesFor 0, 'seconds'
    }
}

The closure is automatically applied to the resolution strategy of every
configuration that’s created by the plugin. This allows the user to
make the necessary customisations while also keeping the invididual
configurations used by the plugin largely hidden.

Closes gh-38