Make exclusion algorithm immune to resolution result’s cache eviction
DefaultResolutionResult uses a cache to speed up access to its root
element. The cache holds the root in memory for 10 seconds after which
time the root will be recreated. The root doesn’t not implement equals
so, for two roots to be equal, they have to be the exact same instance.

ExclusionConfiguringAction uses the root of a DefaultResolutionResult
as the starting point for building the dependency graph. It also calls
ResolutionResult.getAllComponents() to get all of the components in the
result, including the root. It relies on the equality of the root
found in all components and the root returned from
ResolutionResult.getRoot() as it determines what should and should not
be excluded. If cache eviction occurs between access to
all the result’s components and access to its root, the two roots will
not be equal and the search for unexcluded dependencies will end after
the very first iteration.

This commit updates ExclusionConfiguringAction so that it only access
the root of the resolution result and only does so once. This root is
used to build the dependency graph and all the dependencies are then
collected by walking the graph. By accessing the resolution result’s
root only once, ExclusionConfiguringAction is guaranteed to be working
with the same root instances and can rely on their equality.

See gh-39