Break cycles in the dependency graph without pruning too aggressively
Previously, cycles in the dependency graph were broken by not
considering a node’s children if the node had already been seen. This
did break any cycles but it also meant that any node that was seen more
than once would have its descendants cut off for all but its first
appearance in the graph. The missing descendants in the graph could lead
to exclusions being applied too aggressively.

This commit updates the dependency graph to only stop processing the
graph if the node being processed is also one of the node’s ancestors.

Closes gh-47