Align plugin’s exclusion behaviour more closely with Maven’s
Prior to this commit, an exclusion that was declared in a dependency’s
pom would be applied as if it had been declared in dependency
management. For example, if commons-logging was excluded in a dependency
declaration for spring-core, commons-logging would be excluded for
every spring-core dependency, rather than being isolated to the part
of the dependency graph where the exclusion was declared.

This commit updates the exclusion handling logic to only apply an
exclusion from a dependency’s pom to the part of the graph beneath
where it was declared. This, however, isn’t sufficient to align the
behaviour with Maven’s.

When there are conflicts in the graph (multiple occurrences of the same
dependency), Maven resolves them by choosing the occurrence that is
nearest to the root of the graph and removing all other occurrences of
the dependency. This is important in the context of exclusions as it
means that the exclusions declared on the occurrence that’s nearest to
the root will be the ones that take effect.

This commit also updates the exclusion processing to perform the same
conflict resolution as Maven to prune the graph. The graph is then
walked to collect the dependencies that should not be excluded.

Closes gh-50