Avoid using API that’s only available in Java 8 or Groovy 1.x
Groovy 1.x enhances collection to add a sort(Comparator) method. Java 8
has added a default sort(Comparator) method to List which clashes with
Groovy’s enhancement. To avoid this clash, Groovy 2.x has removed its
sort enhancements from Collection. This means that using 
list.sort(Comparator) with Groovy 2.x on Java 7 will fail due to a
missing method. This is a problem as the plugin supports both Gradle
1.x (Groovy 1.8) and Gradle 2.x (Groovy 2.4) and Java 7 and Java 8.

This commit replaces the use of list.sort(Comparator) with a call to
Collections.sort(list, comparator). This ensures that the plugin
functions on all four combinations of Java 7 or 8 and Gradle 1 or 2.