Cope with breaking API changes in Gradle 2.5
Gradle 2.5 has removed a number of repackaged Maven dependencies that
the plugin relied upon. To remain compatible with earlier versions of
Gradle, the plugin cannot use Maven dependencies in their unpackaged
form due to, I believe, a clash with Maven 2 classes that Gradle also
has on the classpath.

This commit updates the plugin to repackage its own copy of the Maven
classes that it depends upon. Sadly, this increases the size of the
plugin unnecessarily for users of Gradle 2.4 and earlier, but it is
the only way to provide a single version of the plugin that is
compatible with versions of Gradle from before and after the change.

Closes gh-54