Allow exclusions to be used alongside failOnVersionConflict()
The process of applying Maven-style exclusions to a configuration
requires access to the poms of all of the configuration’s dependencies
so that it can figure out what to exclude. This requires a copy of
the configuration to be resolved so that the excludes can be applied to
the original configuration before it is resolved.

Previously, a complete copy of the configuration was taken. This copy
also includes any resolution strategies. This is problematic if
failOnVersionConflict has been enabled as the exclusions, that have not
yet been configured, may solve a version conflict that will now stop
the exclusions from being applied.

This commit changes the code that copies the configuration to create a
lighter weight copy. Now, a new configuration is created and only the
dependencies from the original configuration are copied into it. This
configuration copy can then be resolved without be affected by the
failOnVersionConflict setting of the original configuration.

Closes gh-61