Improve the performance of applying Maven exclusion semantics
Previously, a category was used to wall the tree of resolved
components and form a dependency graph from them. Unfortunately, while
quite elegant this approach doesn’t perform satisfactorily. In one,
quite complex reproduction, it was taking approximately 1000ms to create
the graph. Dropping the use of the category has reduced the time taken
to 50ms or less for the same tree.

Closes gh-65