Apply dependency management resolution strategy to all detached configs
Support for a dependency management-specific resolution strategy was
added in 8f59648. This resolution strategy is applied to any
configuration that’s created by the plugin for resolving imported boms.
It is not, however, applied to various other configurations that are
used by the plugin. The configurations that are used when apply Maven
exclusions are one example.

The fact that the resolution strategy isn’t applied to all
detached configurations that are created by the plugin has proven to
be problematic. If a user configures the main configurations to
disable caching this will have no effect. The problem is that the
logic for applying Maven exclusions creates a detached configuration
and copies a main configuration’s dependencies into it. When this
detached configuration is resolved, it’s done so with the default
caching rules. When the main configuration that contains the same
dependency is resolved, the custom caching configuration is ignored.
Presumably, this happens because the dependency has already been
resolved once in this build and Gradle doesn’t want different parts of
the build to use different snapshots.

This commit updates the plugin to introduce a new central location,
DependencyManagementConfigurationContainer, for creating detached
configurations. The dependency management resolution strategy is
apply to every configuration that’s created by this configuration
container, thereby ensuring that things like the caching configuration
are applied consistently.

Closes gh-74