Fail fast when resolution of imported Maven boms fails
Previously, the failure would go unnoticed and the plugin would limp
along until it failed with an unhelpful error message stating that
a configuration cannot be changed after it has been resolved. This was
unhelpful as it masked the underlying cause.

This commit updates the plugin to throw a GradleException when the
attempt to resolve the configuration that contains the imported bom
dependencies fails. This causes the build to fail fast, and 



Closes gh-81