Apply license plugin to build
This patch applies and configures `hierynomus/license-gradle-plugin`
to keep file headers up to date. The settings can be changed to make
the build fail if a header is missing or non-conformant. For now, all
file headers have been kept untouched.

Run `gradle license` to obtain a report of non-conformant files.
Run `gradle licenseformat` to apply the license template to all files.

Please review the settings found in the `license` block.