Apply additional metadata on JAR manifest.
This patch adds useful metadata to the JAR manifest, such as JDK,
user, buidl date and time. The SCM hash was ommitted as the
`versioning` plugin is not supported by Gradle 1.12. Please consider
a different method to obtain such value _or_ migrate the build
to Gradle 2.x if possible.

Fixes #85