Don’t apply dependency management to dependencies with dynamic versions
Previously, the plugin would apply dependency management to a dependency
with a dynamic version. This doesn’t make much sense as the version is
then no longer dynamic. This commit updates the plugin so that
dependency management is not applied to dependencies with a dynamic
version. A version if deemed to be dynamic if it:

- Ends with +
- Starts with latest.
- Is a version range

Closes gh-77