Apply user-provided resolution strategy after the managed version strategy
Previously, if a user provided a dependency management resolution
strategy it would be called before the internal strategy that applies
managed versions. This proved to be problematic for dependencies when
the user-provided resolution strategy was setting the version as it
would, potentially, then be overridden by a managed version.

This commit reverses the order in which the resolution strategies are
configured so that a user-provided resolution strategy will run after
the strategy that applies managed versions, thereby allowing the user’s
strategy to take precedence.

See gh-98