Consider unresolvable dependencies when determining exclusions
Previously, any dependency that could not be resolved while exclusions
were being calculated would cause a failure. This meant that an
exclusion could not be used to prevent a dependency that could not be
resolved from causing a resolution failure.

This commit updates ExclusionConfiguringAction to give special
consideration to dependency that cannot be resolved.

Firstly, when the configuration copy that is used to resolve
dependencies and their exclusions fails to resolve, the resolution
result is analysed. If the failure was only due to dependencies which
we already know should be excluded it is ignored.

Secondly, when collecting all of the configuration’s dependencies, 
unresolvable dependencies are considered. Unresolvable dependencies are
then also considered when determining which dependencies should be
included, i.e. which have not been excluded.

Closes gh-118