Tolerate all errors and ModelBuildingException in EffectiveModelBuilder
Previously, in the event of any errors or a ModelBuildingException,
EffectiveModelBuilder would return null. This could then result in a
NullPointerException in calling code.

Rather than requiring callers to deal with null, this commit changes
EffectiveModelBuilder to report any errors that have occurred and then
return the effective model anyway. This is a reasonable approach to take
a such faulty poms do not generally cause Maven builds to fail.

Closes gh-144