Work around limitations in Groovy's type coercion
Groovy appears to only perform type coercion at runtime. This is
problematic when generics are used as the generic information is lost
during compilation. As a result, the desired type coercion is not
performed at runtime as there's no way of knowing what is the desired
type. This was leading to methods in the plugin's DSL that take a
Map<String, String> being called with a map that contained
GStrings. A GString is not a String so a ClassCastException would
result when trying to access entries in the map.

This commit updates the implementations of the DSL to treat any
Map<String, String> as potentially containing any implementation of
CharSequence. The API itself has not been changed to avoid breaking
binary compatibility.

Closes gh-148