Apply dependency management to transitive deps with dynamic versions
Previously, any dependency with a dynamic version was skipped. This
meant that a transitive dependency that was depended upon using
a version range would not have its version managed.

The skipping was introduced for a case where a user had expressly used
a dynamic version, either in directly in their build.gradle or because
they were using the Versions plugin. This case can still be satisfied
by only skipped direct dependencies that use a dynamic version.

This commit updates VersionConfiguringAction so that dependency
management is applied to transitive dependencies with dynamic
versions, thereby allowing dependency management to take control over
the version of any such dependencies.

Closes gh-151