Stop pom customization from breaking custom conf-to-scope mappings
Accessing a MavenResolver's pom has an unwanted side-effect of
configuring it with new Conf2ScopeMappings that are a snapshot of the
project's current Conf2ScopeMappings. As a result, any subsequent
changes to the project's Conf2ScopeMappings are lost and are then not
available during pom generation. This behaviour has been reported
to the Gradle team [1] and, it would appear, accepted as a bug.

This commit works around the problem by deferring the configuration of
any pom customization until after the project has been evaluated. This
gives as much time as possible for other plugins and the user's own
configuration to configure Conf2ScopeMappings before the snapshot is
taken.

Closes gh-149

[1] https://github.com/gradle/gradle/issues/1373