Restore compatibility with the Versions plugin
The changes made in 03237ef were intended to allow the versions of
transitive dependencies with dynamic versions to be managed, while
ignoring direct dependencies with dynamic versions. For the most part
this worked as intended, but it did not work when the Versions plugin
[1] was used.

The Versions plugin takes a copy of a configuration and modifies its
dependencies such that each then has a dynamic version. When the
configuration is copied, the before resolve action that was used by
VersionConfiguringAction was not copied over (this may well be a bug
in Gradle [2]). This means that the action incorrectly believed that
the configuration had no direct dependencies and, as a result, it
applied dependency management to all of the configuration's
dependencies. This prevented the versions plugin from identifying
dependencies for which a newer version is available.

This commit updates VersionConfiguringAction so that it no longer uses
a before resolve action. Instead, it lazily gets the coordinates of
all of the configuration's dependencies the first time it is called.
This allows the direct dependencies to be identified correctly, even
when its a copy of a configuration that it being resolved. Note that
this will only work if the dependencies in the configuration copy
have the same groups and names as the dependencies in the original
configuration as is the case with the Versions plugin.

Closes gh-155

[1] https://github.com/ben-manes/gradle-versions-plugin
[2] https://github.com/gradle/gradle/issues/1567