Allow DSL's configurations to be called with String or Configuration
Previously, the configurations "method" that's handled by Groovy's
method missing infrastructure expected the arguments to be instances
of DependencyManagementConfigurer and ignored anything that was not.

This commit makes the DSL more flexible by also allowing
configurations to be called with a Configuration (used to create a
Configurer) or a String (used look up a Configuration and then
create a Configurer).

Closes gh-174