Reflect version overrides in generated poms
Previously, when a bom was imported and a version property in that
pom was overridden, the override was not reflecting in the dependency
management configuration in the generated Maven pom.

This commit updates the dependency management configuration logic to
consider version properties and add overriding <dependency> nodes in
the generated pom's <dependencyManagement> that reflect the overridden
version.

Closes gh-184