Fix handling of overridden transitive bom import with new dependency
Previously, if the version used for a transitive bom import was
overridden and the override results in the bom containing additional
dependency management, pom customization would fail with a null
pointer exception. A concrete example of this failure is an import
of Spring Boot's spring-boot-dependencies bom and an override of the
spring.version property to use Spring Framework 5.0.x. The new
spring-context-indexer module in Framework 5 was not handled properly
and a NPE occurred.

This commit updates the plugin to correctly handle the case where
a dependency is managed in the overridden version of the bom but not
in the original version.

Closes gh-198