Add imported boms to generated pom in correct order and fix overrides
Previously, imported boms were imported into a generated pom in the
order in which they're imported in build.gradle. This causes a problem
Where the imported boms provided dependency management for the same
dependency, this led to the version being wrong in the Maven project
as Maven uses first-wins whereas this plugin uses last wins. A
further problem where dependency management overlapped between two
boms was that an explicit override would be added to the generated
pom. This override is unnecessary as the dependency management in one
imported bom will naturally override the other with an explicit
declaration.

This commit reverses the order in which imported boms are imported
into the generated pom. It also refines the logic that's used to
determine whether or not an explicit dependency management override
is needed so that an override is not generated when one bom is
overriding another.

Closes gh-204