Ignore dependency management with a classifier in imported boms
Previously, when importing a bom, dependency management with a
classifier was treated exactly the same as dependency management
without a classifier. This was done despite the fact that Gradle's
dependency resolution APIs do not make it possible to perform
classifier-specific dependency management.

In addition to it being pointless as the classifier-specific
dependency management could not be applied, importing the managed
version for a dependency with a classifier could be harmful. If
dependency management was provided for the same group id and artifact
id both with and without a classifier then, in the unlikely event of
the versions being different, the dependency management with a
classifier could overwrite the dependency management without a
classifier.

To avoid the above-described problem, this commit updates the plugin
so that it ignores dependency management with a classifier when
importing a bom.

Closes gh-218