Delegate missing methods to project in DSL extension
Previously, the dependency management DSL extension's handling of
missing methods would look for a configuration with a name the
same as that of the missing method. If it did not find one an
exception would be thrown. This prevented methods on the project
being called from within the extension.

This commit updates the missing method handling of the extension to
address the above-described problem. When a matching configuration
cannot be found, the method invocation is delegated to the project.

Closes gh-255