Cache POM exclusions in Gradle daemon
In many of our spring projects dependency resolution is the slowest link
in the edit-compile-test cycle, ranging between ~1.5 to ~6 seconds in
the worst case.

Instrumenting the slowest of our multi-project builds I noticed that
over 1,700 Pom files are parsed during dependency resolution, with some
POMS being parsed many times over.

This change introduces a static POM exclusion cache, which reduces
dependency resolution to less than a second on subsequent builds in that
 same project.