Ignore platform dependencies when determining exclusions
Previously, when a dependency with Gradle module metadata had a
transitive dependency on a platform, the platform and its
graph of dependency contraints would be treated as standard
dependencies when identifying the dependencies that should be
included. This could result in a dependency being including
incorrectly if the dependency contraints created a path to it that
was not covered by an exclusion.

This commit updates the logic that determines the included
dependencies to stop walking down the dependency graph when a node
in the graph is a platform dependency.

Fixes gh-310