Avoid excluding parent poms
In 3d05983, the fix for gh-310, an attempt was made to prevent
dependencies in a transitive platform dependency from stoping
exclusions from being applied correctly. Unfortunately, it went
to far which led to a dependency's parent pom being excluded. This
could cause resolution failures if that parent contained
dependency management that was required for the child's dependencies
to resolve correctly.

This commit reworks that changes so that the platform dependency
itself is added to the included nodes but to then stop processing.
This prevents the platform dependency itself from being excluded but
ensures that any of its dependencies due not influence the
application of exclusions.

Fixes gh-360