Fix platform constraint processing to avoid accidental exclude
Previously, a constraint in a platform that applied to a transitive
dependency of the platform would prevent subsequent processing of
that dependency. This resulted in the dependency being
unintentionally excluded.

This commit updates the configuring of exclusions to only track a
dependency is seen when it's an actual dependency and not a
constraint.

Closes gh-368