Consider changes to a dependency's target when managing its version
Previously, dependency management was determined used the requested
selector. This meant that any changes made to the target selector
were ignored. This would cause a problem if an earlier resolution
strategy had changed the group ID or artifact ID of the dependency
as the managed version would be determined using the original
coordinates not the updated coordinates.

This commit applies managed versions using the target selector,
thereby ensuring that any earlier changes to the target are taken
into account when determining the managed version.

Fixes gh-383